% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_set_spatial_buffer.R
\name{create_set_spatial_buffer}
\alias{create_set_spatial_buffer}
\title{Find sets of units that are separated by a minimum distance (spatial buffer).}
\usage{
create_set_spatial_buffer(distance_matrix, threshold, sims = 1000,
  probability_weights = NULL, exclude_set = NULL, ...)
}
\arguments{
\item{distance_matrix}{A distance matrix (see create_distance_matrix()).}

\item{threshold}{The minimum distance one unit should be from another (expressed in same units as distance matrix)}

\item{sims}{The number of iterations to run in order to find sets.}

\item{exclude_set}{A list of vectors (can be created with find_proximate_units()), each element of the list corresponds to each unit, and contains a vector of all of the units to which that unit is proximate.}
}
\value{
A list of admissable sets of units that meet the spatial criterion.
}
\description{
Find sets of units that are separated by a minimum distance (spatial buffer).
}

